pyproject.toml
[project]
name = "Travel_Agent"
version = "0.1.0"
description = "Langgraph Travel Agent"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.23",
    "langchain-community>=0.3.21",
    "langchain-openai>=0.3.14",
    "langgraph>=0.3.31",
    "langgraph-cli[inmem]>=0.2.5",
    "tavily-python>=0.5.4",
]
